Perform expensive subset size calculation only once
Calculating the size (as well as isEmpty) on a subset of a SortedSet is a somewhat expensive operation that involves walking the tree of the parent set (since a subset object is merely a thin wrapper around the full set). It is not O(1) like you might expect.

This removes the isEmpty() call, so we walk the tree one fewer times.